+++ src/Components/Dropdowns/Composed/TypeAheadDropDown.tsx
@@ -1,12 +1,5 @@
 // Libraries
-import React, {
-  ChangeEvent,
-  FC,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from 'react'
+import React, {ChangeEvent, FC, useEffect, useMemo, useState} from 'react'
 import classnames from 'classnames'
 import {Dropdown} from '../.'
 import {MenuStatus} from '../Dropdown'
@@ -19,7 +12,9 @@ import {
 
 import {Input} from '../../Inputs/Input'
 import {DropdownHeader} from '../DropdownHeader'
-
+import {FixedSizeList} from 'react-window'
+import '../ScrollBarStyles.scss'
+import './TypeAheadDropDownStyles.scss'
 export interface SelectableItem {
   id: string
   name?: string
@@ -34,10 +29,6 @@ interface OwnProps extends StandardFunctionProps {
   testIdSuffix?: string
   /**optional pre-selected option, must match exactly (name and id) an item in the items array */
   selectedOption?: SelectableItem | null
-  /**enables forced searching once dropdown list exceeds largeListSearch value */
-  largeListSearch?: boolean
-  /**the number of total items in the dropdown list before search is forced */
-  largeListCeiling?: number
   /** which theme to apply */
   menuTheme?: DropdownMenuTheme
   buttonTestId?: string
@@ -49,20 +40,10 @@ interface OwnProps extends StandardFunctionProps {
   /** status state: default, loading, or disabled */
   status?: ComponentStatus
 }
-const isBlank = (pString: string | undefined): boolean =>
-  // Checks for falsiness or a non-white space character
-  !pString || !/[^\s]+/.test(pString)
 
-const getValueWithBackup = (
-  val: string | undefined,
-  backup: string
-): string => {
-  if (isBlank(val)) {
-    return backup
-  }
-  return val as string
-}
 const enCollator = new Intl.Collator('en-us')
+
+const LIST_ITEM_HEIGHT = 33
 
 export const TypeAheadDropDown: FC<OwnProps> = ({
   id,
@@ -70,16 +51,10 @@ export const TypeAheadDropDown: FC<OwnProps> = ({
   items,
   onSelect,
   testID,
-  placeholderText,
-  testIdSuffix = 'typeAhead',
+  placeholderText = 'Select a Value',
   selectedOption = null,
   className,
-  largeListSearch = false,
-  largeListCeiling = 0,
   menuTheme = DropdownMenuTheme.Onyx,
-  buttonTestId = 'type-ahead-dropdown--button',
-  menuTestID = 'type-ahead-dropdown--menu',
-  itemTestIdPrefix = 'type-ahead-dropdown--item',
   sortNames = true,
   defaultNameText = '',
   status = ComponentStatus.Default,
@@ -94,28 +69,44 @@ export const TypeAheadDropDown: FC<OwnProps> = ({
 
   const [selectIndex, setSelectIndex] = useState(-1)
   const [queryResults, setQueryResults] = useState(items)
-  const [menuOpen, setMenuOpen] = useState<MenuStatus>(MenuStatus.Closed)
-
+  const [menuStatus, setMenuStatus] = useState<MenuStatus>(MenuStatus.Closed)
+  const [userHasTyped, setUserHasTyped] = useState(false)
   const [selectedItem, setSelectedItem] = useState<SelectableItem | null>(
     selectedOption
   )
 
-  let initialTypedValue = ''
+  const listRef = React.createRef<FixedSizeList<SelectableItem[]>>()
+
+  let initialInputValue = ''
 
   if (selectedOption) {
-    initialTypedValue = getValueWithBackup(selectedOption.name, defaultNameText)
+    initialInputValue = selectedOption.name || defaultNameText
   }
 
-  const [typedValue, setTypedValue] = useState<string>(initialTypedValue)
-
-  const largeListValidation =
-    largeListSearch && queryResults.length > largeListCeiling
+  const [inputValue, setInputValue] = useState<string>(initialInputValue)
+  const [backupSelectionValue, setBackupSelectionValue] = useState<string>(
+    initialInputValue
+  )
 
   useEffect(() => {
-    if (typedValue.length > 0) {
+    // If selectedOption is invalid (not in the list of items), then we want to clear the input value
+    if (
+      selectedOption &&
+      !items.some(item => item.name === selectedOption.name)
+    ) {
+      setInputValue('')
+      return
+    }
+    setInputValue(selectedOption?.name || defaultNameText)
+    setBackupSelectionValue(selectedOption?.name || defaultNameText)
+  }, [selectedOption])
+
+  useEffect(() => {
+    // only filter the list to show the results if the user has typed something
+    if (inputValue.length > 0 && userHasTyped) {
       const result = items.filter(val => {
         const name = val?.name || ''
-        return name.toLowerCase().includes(typedValue.toLowerCase())
+        return name.toLowerCase().includes(inputValue.toLowerCase())
       })
 
       // always reset the selectIndex when doing filtering;  because
@@ -126,62 +117,38 @@ export const TypeAheadDropDown: FC<OwnProps> = ({
     } else {
       setQueryResults(items)
     }
-  }, [items, typedValue])
-
-  /**
-   *  using a ref to hold an instance variable:  what was last typed,
-   *  because without this 'click to select' doesn't work.
-   *  (you click to select, which clicks out of the dropdown, so then the dropdown sets
-   * the text to what was last selected.  this works fine for a class component,
-   * but here it uses the stale state of what was previously selected.
-   * this way, what was selected is saved in the ref.)
-   */
-  const backupValue = useRef<string>(initialTypedValue)
+  }, [items, inputValue])
 
   const itemNames = useMemo(() => items.map(item => item.name?.toLowerCase()), [
     items.length,
   ])
 
-  /**
-   *  filter the selections/options based on the search string: needle
-   * if the needle is empty; then there is nothing to filter; so return everything
-   */
-  const doFilter = (needle: string) => {
-    // if there is no value, set the queryResults to everything
-    // and set the typedValue to nothing (zero it out)
-    // reset the selectIndex too
-    if (!needle) {
+  const filter = (filterStr: string) => {
+    if (!filterStr) {
       setQueryResults(items)
-      setTypedValue('')
+      setInputValue('')
       setSelectIndex(-1)
     } else {
       const result = items.filter(val => {
         const name = val?.name || ''
-        return name.toLowerCase().includes(needle.toLowerCase())
+        return name.toLowerCase().includes(filterStr.toLowerCase())
       })
 
-      // always reset the selectIndex when doing filtering;  because
-      // if it had a value, and then they type, the queryResults changes
-      // so need to reset
       setQueryResults(result)
-      setTypedValue(needle)
-      setMenuOpen(MenuStatus.Open)
+      setUserHasTyped(true)
+      setInputValue(filterStr)
       setSelectIndex(-1)
     }
   }
 
-  const clear = () => {
-    doSelection(null)
-    doFilter('')
+  const handleInputChange = (event: ChangeEvent<HTMLInputElement>) => {
+    const filterStr = event?.target?.value
+    filter(filterStr)
   }
 
-  const filterVals = (event: ChangeEvent<HTMLInputElement>) => {
-    const needle = event?.target?.value
-    doFilter(needle)
-  }
-
-  if (!placeholderText) {
-    placeholderText = 'Select a Value'
+  const onClear = () => {
+    selectItem(null)
+    filter('')
   }
 
   const setTypedValueToSelectedName = (backupName?: string) => {
@@ -189,19 +156,17 @@ export const TypeAheadDropDown: FC<OwnProps> = ({
     if (!selectedName) {
       selectedName = ''
     }
-    setTypedValue(selectedName)
+    setInputValue(selectedName)
   }
 
-  const maybeSelectNextItem = (
+  const handleKeyboardUpDown = (
     event: React.KeyboardEvent<HTMLInputElement>
   ) => {
     let newIndex = -1
 
-    if (event.keyCode === 40) {
-      // down arrow
+    if (event.key === 'ArrowDown') {
       newIndex = selectIndex + 1
-    } else if (event.keyCode === 38) {
-      // up arrow
+    } else if (event.key === 'ArrowUp') {
       newIndex = selectIndex - 1
     }
 
@@ -213,27 +178,24 @@ export const TypeAheadDropDown: FC<OwnProps> = ({
       return
     }
 
-    if (event.keyCode === 13) {
-      // return/enter key
+    if (event.key === 'Enter') {
       // reset the selectIndex to -1, & close the menu:
-
       if (numItems && selectIndex >= 0 && selectIndex < numItems) {
-        // they used the arrows; just pressed return
-        doSelection(queryResults[selectIndex], true)
+        selectItem(queryResults[selectIndex])
       } else {
         // the person could have been typing and pressed return, need to
         // make sure the value in the input field is real/legal:
 
         // but:  if the value they typed is LEGAL (in the list/dropdown values), set it;
         // else: reset to the previous real/legal value:
-        const foundIndex = itemNames.indexOf(typedValue.toLowerCase())
+        const foundIndex = itemNames.indexOf(inputValue.toLowerCase())
 
         if (foundIndex >= 0) {
           // is a real legal value
-          doSelection(items[foundIndex], true)
+          selectItem(items[foundIndex])
         } else {
           setTypedValueToSelectedName()
-          setMenuOpen(MenuStatus.Closed)
+          setMenuStatus(MenuStatus.Closed)
           setSelectIndex(-1)
         }
       }
@@ -246,31 +208,34 @@ export const TypeAheadDropDown: FC<OwnProps> = ({
    * */
   const getDisplayName = (item: SelectableItem | null): string => {
     if (item && item.id) {
-      return getValueWithBackup(item.name, defaultNameText)
+      return item.name ?? defaultNameText
     }
     return ''
   }
 
-  const doSelection = (item: SelectableItem | null, closeMenuNow?: boolean) => {
+  const selectItem = (item: SelectableItem | null) => {
     setSelectedItem(item)
     const actualName = getDisplayName(item)
-    setTypedValue(actualName)
-    backupValue.current = actualName
+    setInputValue(actualName)
     setSelectIndex(-1)
 
-    if (closeMenuNow) {
-      setMenuOpen(MenuStatus.Closed)
+    // selectItem is called from onClear, we don't close the menu
+    if (item !== null) {
+      setMenuStatus(MenuStatus.Closed)
+      setBackupSelectionValue(actualName)
     }
     onSelect(item)
   }
 
-  const onClickAwayHere = () => {
+  const onClickOutside = () => {
     //  reset to the selected value; if the user typed in
     //  something not allowed it will go back to the last selected value:
-    setTypedValueToSelectedName(backupValue.current)
+    setTypedValueToSelectedName(backupSelectionValue)
+    setQueryResults(items)
+    setMenuStatus(MenuStatus.Closed)
+    setUserHasTyped(false)
+    setSelectIndex(-1)
   }
-
-  const dropdownStatus = items.length === 0 ? ComponentStatus.Disabled : status
 
   const placeText =
     status === ComponentStatus.Loading ? 'Loading...' : placeholderText
@@ -281,91 +246,114 @@ export const TypeAheadDropDown: FC<OwnProps> = ({
     }
   }
 
-  // do rendering:
   const inputComponent = (
     <Input
       placeholder={placeText}
-      onChange={filterVals}
-      value={typedValue}
-      onKeyDown={maybeSelectNextItem}
-      testID={`dropdown-input-typeAhead--${testIdSuffix}`}
-      onClear={clear}
+      onChange={handleInputChange}
+      value={inputValue}
+      onKeyDown={handleKeyboardUpDown}
+      testID={`${testID}--typeAhead-input`}
+      onClear={onClear}
       status={status}
       onFocus={selectAllTextInInput}
     />
   )
 
-  const props: any = {id, style, className, menuOpen}
+  const props: any = {id, style, className, menuOpen: menuStatus}
 
-  const largeListValidationText =
-    typedValue.length >= 1
-      ? 'There are still too many results. Please input more characters.'
-      : 'Please input a character to start seeing results.'
+  const toggleMenu = (event: any) => {
+    if (
+      (event.target.className === 'cf-input-field' ||
+        event.target.className.includes('cf-dismiss-button')) &&
+      menuStatus === MenuStatus.Open
+    ) {
+      return
+    }
+    // toggle the menu:
+    if (menuStatus === MenuStatus.Closed) {
+      setMenuStatus(MenuStatus.Open)
+    } else {
+      setMenuStatus(MenuStatus.Closed)
+    }
+  }
+
+  const getSelectedItemIndex = (): number => {
+    if (selectedItem) {
+      return items.findIndex(item => item.id === selectedItem.id)
+    }
+    return 0
+  }
 
   return (
     <Dropdown
       {...props}
-      testID={testID || `typeAhead-dropdown--${testIdSuffix}`}
-      onClickAway={onClickAwayHere}
+      testID={testID}
+      onClickAway={onClickOutside}
       disableAutoFocus
-      button={(active, onClick) => (
+      button={active => (
         <DropdownHeader
           active={active}
-          onClick={onClick}
-          testID={buttonTestId}
-          status={dropdownStatus}
+          onClick={toggleMenu}
+          testID={`typeAhead-dropdown--button`}
         >
           {inputComponent}
         </DropdownHeader>
       )}
-      menu={onCollapse => (
-        <Dropdown.Menu
-          testID={menuTestID}
-          onCollapse={onCollapse}
-          theme={menuTheme}
-        >
-          {largeListValidation ? (
-            <Dropdown.Item
-              key="nada-no-values-in-filter"
-              testID="nothing-in-filter-typeAhead"
-              disabled={true}
-              wrapText={true}
+      menu={() => (
+        <Dropdown.Menu testID={`${testID}-dropdown-menu`} theme={menuTheme}>
+          {queryResults && queryResults.length > 0 ? (
+            <FixedSizeList
+              height={
+                queryResults.length * LIST_ITEM_HEIGHT > 150
+                  ? 150
+                  : queryResults.length * LIST_ITEM_HEIGHT
+              }
+              itemCount={queryResults.length}
+              itemSize={33}
+              width={'100%'}
+              layout="vertical"
+              itemData={queryResults}
+              className="menu-dropdown"
+              ref={listRef}
+              initialScrollOffset={getSelectedItemIndex() * LIST_ITEM_HEIGHT}
             >
-              {largeListValidationText}
-            </Dropdown.Item>
-          ) : (
-            queryResults.map((value, index) => {
-              // add the 'active' class to highlight when arrowing; like a hover
-              const classN = classnames({
-                active: index === selectIndex,
-              })
+              {({data, index, style}) => {
+                const value = data[index]
+                // add the 'active' class to highlight when arrowing; like a hover
+                const classN = classnames('menu-dropdown-typeahead-item', {
+                  active: index === selectIndex,
+                })
 
-              return (
-                <Dropdown.Item
-                  key={value.id}
-                  id={value.id}
-                  value={value}
-                  onClick={doSelection}
-                  selected={value.id === selectedItem?.id}
-                  testID={`${itemTestIdPrefix}-${value.id}`}
-                  className={classN}
-                >
-                  {value.name || defaultNameText}
-                </Dropdown.Item>
-              )
-            })
-          )}
-          {!queryResults || queryResults.length === 0 ? (
+                return (
+                  <div key={value.id} style={style}>
+                    <Dropdown.Item
+                      id={value.id.toString()}
+                      value={value}
+                      onClick={() => selectItem(value)}
+                      selected={value.id === selectedItem?.id}
+                      testID={`typeAhead-dropdown--item`}
+                      className={classN}
+                    >
+                      {value.name || defaultNameText}
+                    </Dropdown.Item>
+                  </div>
+                )
+              }}
+            </FixedSizeList>
+          ) : (
             <Dropdown.Item
-              key="nada-no-values-in-filter"
+              key="no-values-in-filter"
               testID="nothing-in-filter-typeAhead"
               disabled={true}
             >
-              {`no matches for ${typedValue}`}
+              {inputValue.length > 0
+                ? `no matches for ${inputValue}`
+                : 'No results'}
             </Dropdown.Item>
-          ) : null}
+          )}
         </Dropdown.Menu>
       )}
+      menuOpen={menuStatus}
     />
   )
 }